Implements MPC for Autonomous Driving
This is a "Finally-I-Can-Complete-a-Lapse-Yaayy-Solution".

Need to understand impact of each cost function thoroughly and experiment
speeding up. I am not sure I fnd it easy to understand the impact of
each cost function.

TODO:
Implement a way to pass cost weights through command line for thorough
experiments.

Assumption:
Latency value between mpc solver and simulator is taken as 100ms. I dont
plan to change this as this is not something I want to get accurate for
now.